requirements.txt
opencv-python
tqdm
transformers
pytorch
numpy
pandas
SimpleITK
sklearn
random
copy
radiomics